# specs/agent-portal/NDAI-27811-init-api-call/contracts/portal-application-prefill.openapi.yaml
openapi: 3.0.3
info:
  title: Agent Portal - Application Prefill API
  version: 0.1.0
  description: >
    Internal Agent Portal API used to fetch and bind application data into UI view models
    (Basic Information, Drivers, Vehicles). This API is a server-side boundary and must
    not expose secrets or raw upstream payloads unnecessarily.
servers:
  - url: https://{host}
    variables:
      host:
        default: quote.uat.novo.us
paths:
  /api/application-prefill:
    post:
      summary: Prefill quote workflow view models from encoded payload
      operationId: applicationPrefill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - encodedPayload
              properties:
                encodedPayload:
                  type: string
                  description: Encoded payload extracted from inbound link (base64/base64url).
                currentDraftId:
                  type: string
                  description: Current local draft identifier (used to detect conflicts).
      responses:
        "200":
          description: Prefill view models ready to bind into QuoteContext
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationPrefillResponse"
        "400":
          description: Invalid or malformed encoded payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied (no retry, no data)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited (client should surface retry behavior per policy)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    ApplicationPrefillResponse:
      type: object
      additionalProperties: false
      required:
        - quoteId
        - applicationId
        - replaceLocalDraft
        - viewModels
      properties:
        quoteId:
          type: string
        applicationId:
          type: string
        replaceLocalDraft:
          type: boolean
          description: Always true when inbound application differs from local draft.
        viewModels:
          type: object
          additionalProperties: false
          required:
            - basicInfo
            - drivers
            - vehicles
          properties:
            basicInfo:
              type: object
              description: Basic information view model (ClientInfoFormValues shape).
              additionalProperties: true
            drivers:
              type: array
              items:
                type: object
                description: Driver view model (Driver[] shape).
                additionalProperties: true
            vehicles:
              type: array
              items:
                type: object
                description: Vehicle view model (Vehicle[] shape).
                additionalProperties: true
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable error code (no PII).
        message:
          type: string
          description: User-friendly message (no secrets, no PII).